pyproject.toml
[project]
name = "EvoScientist"
version = "0.1.0"
description = "EvoScientist: Towards Self-Evolving AI Scientists for End-to-End Scientific Discovery"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Xi Zhang" },
]
maintainers = [
    { name = "Xi Zhang" },
]
keywords = ["ai-scientific", "scientific-discovery", "self-evolving", "ai-scientists", "end-to-end"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "deepagents>=0.3.6",
    "langchain>=1.2",
    "langchain-anthropic>=1.3",
    "anthropic>=0.76",
    "tavily-python>=0.7",
    "pyyaml>=6.0",
    "rich>=14.0",
    "prompt-toolkit>=3.0",
    "python-dotenv>=1.0",
    "langgraph-cli[inmem]>=0.4",
    "httpx>=0.27",
    "markdownify>=0.14",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.urls]
"Homepage" = "https://github.com/EvoScientist/EvoScientist"
"Bug Tracker" = "https://github.com/EvoScientist/EvoScientist/issues"
"Documentation" = "https://github.com/EvoScientist/EvoScientist#readme"

[project.scripts]
evoscientist = "EvoScientist.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["EvoScientist*"]

[tool.setuptools.package-data]
EvoScientist = ["subagent.yaml"]
